pyproject.toml
[project]
name = "disaster-routing"
version = "0.1.0"
description = "Disaster Routing for Elastic Optical Network experiments"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "hydra-joblib-launcher>=1.2.0",
    "lxml>=6.0.0",
    "matplotlib>=3.10.3",
    "maturin>=1.9.3",
    "networkx>=3.5",
    "numpy>=2.3.0",
    "openpyxl>=3.1.5",
    "pulp>=3.2.1",
    "scipy>=1.15.3",
    "scipy-stubs>=1.15.3.0",
    "types-networkx>=3.5.0.20250610",
]

[tool.pyright]
typeCheckingMode = "recommended"
reportAny = "none"
reportExplicitAny = "none"

[dependency-groups]
color-log = [
    "colorama>=0.4.6",
    "colorlog>=6.9.0",
    "pygments>=2.19.2",
]
dev = [
    "fs>=2.4.16",
    "jq>=1.9.1",
    "pandas>=2.3.1",
    "pandas-stubs>=2.3.0.250703",
    "pre-commit>=4.3.0",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "xlsxwriter>=3.2.5",
]

[tool.uv]
default-groups = ["dev", "color-log"]

[tool.ruff.lint]
select = ["E", "F", "I", "TID"]
